feat: rename dirs modal component and update modal imports
