+++ app/layout.tsx
@@ -2,8 +2,8 @@ import type { Metadata } from "next";
 import { Inter } from "next/font/google";
 import { Providers } from "./providers";
 import "./globals.css";
-import { ModalRender } from "@/components/Modals/ModalRender";
-import { ShadcnModalRender } from "@/components/Modals/ShadcnModalRender";
+import { ModalRender } from "@/components/modals/ModalRender";
+import { ShadcnModalRender } from "@/components/modals/shadcn-modal-render";
 import { AppSidebar } from "@/components/app-sidebar";
 import { NavActions } from "@/components/nav-actions";
 import { Separator } from "@/components/ui/separator";
